+++ OnnxStack.StableDiffusion/Helpers/TensorHelper.cs
@@ -247,15 +247,6 @@ namespace OnnxStack.StableDiffusion.Helpers
         }
 
 
-      
-
-
-
-
-
-
-
-
         /// <summary>
         /// Generate a random Tensor from a normal distribution with mean 0 and variance 1
         /// </summary>
@@ -279,58 +270,5 @@ namespace OnnxStack.StableDiffusion.Helpers
             return latents;
         }
 
-
-        /// <summary>
-        /// Splits the Tensor along axis 0.
-        /// </summary>
-        /// <param name="tensor">The tensor.</param>
-        /// <param name="count">The count.</param>
-        /// <param name="axis">The axis.</param>
-        /// <returns></returns>
-        /// <exception cref="System.NotImplementedException">Only axis 0 is supported</exception>
-        public static DenseTensor<float>[] Split(this DenseTensor<float> tensor, int count, int axis = 0)
-        {
-            if (axis != 0)
-                throw new NotImplementedException("Only axis 0 is supported");
-
-            var dimensions = tensor.Dimensions.ToArray();
-            dimensions[0] /= count;
-
-            var newLength = (int)tensor.Length / count;
-            var results = new DenseTensor<float>[count];
-            for (int i = 0; i < count; i++)
-            {
-                var start = i * newLength;
-                results[i] = new DenseTensor<float>(tensor.Buffer.Slice(start, newLength), dimensions);
-            }
-            return results;
-        }
-
-
-        /// <summary>
-        /// Joins the tensors across the 0 axis.
-        /// </summary>
-        /// <param name="tensors">The tensors.</param>
-        /// <param name="axis">The axis.</param>
-        /// <returns></returns>
-        /// <exception cref="System.NotImplementedException">Only axis 0 is supported</exception>
-        public static DenseTensor<float> Join(this IList<DenseTensor<float>> tensors, int axis = 0)
-        {
-            if (axis != 0)
-                throw new NotImplementedException("Only axis 0 is supported");
-
-            var tensor = tensors.First();
-            var dimensions = tensor.Dimensions.ToArray();
-            dimensions[0] *= tensors.Count;
-
-            var newLength = (int)tensor.Length;
-            var buffer = new float[newLength * tensors.Count].AsMemory();
-            for (int i = 0; i < tensors.Count(); i++)
-            {
-                var start = i * newLength;
-                tensors[i].Buffer.CopyTo(buffer[start..]);
-            }
-            return new DenseTensor<float>(buffer, dimensions);
-        }
     }
 }
